Attempts into bringing soft renderer support.

+++ code/renderercommon/qgl.h
@@ -29,8 +29,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #define MAX_INDICES 4096
 #define VERTEXARRAYSIZE 18360
 extern float *gVertexBuffer;
-extern uint8_t *gColorBuffer;
-extern uint8_t *gColorBuffer255;
+extern float *gColorBuffer;
+extern float *gColorBuffer255;
 extern float *gTexCoordBuffer;
 extern uint16_t *indices;
